+++ python/omfiles/__init__.py
@@ -1,4 +1,4 @@
 from . import types, xarray_backend
-from .omfiles import OmFilePyReader, OmFilePyWriter
+from .omfiles import OmFilePyReader, OmFilePyReaderAsync, OmFilePyWriter
 
-__all__ = ["OmFilePyReader", "OmFilePyWriter", "xarray_backend", "types"]
+__all__ = ["OmFilePyReader", "OmFilePyReaderAsync","OmFilePyWriter", "xarray_backend", "types"]