+++ CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added Changelog
 - Added example for a latitude-longitude plot
+- Added Async Reader
 
 ### Fixed
 
+- Fix type hint for shape property of OmFilePyReader
 - Improved tests to use `pytest` fixtures
 - Fix xarray contained attributes as variables
 - Improve benchmarks slightly